Move side calcs to g_input
