+++ src/g_game.c
@@ -153,7 +153,7 @@ wbstartstruct_t wminfo;               // parms for world map / intermission
 boolean         haswolflevels = false;// jff 4/18/98 wolf levels present
 byte            *savebuffer;
 boolean         autorun = false;      // always running?          // phares
-static boolean  autostrafe50;
+boolean         autostrafe50;
 boolean         novert = false;
 boolean         mouselook = false;
 boolean         padlook = false;
@@ -193,11 +193,10 @@ static boolean mouse_y_invert;
 #define QUICKREVERSE 32768 // 180 degree reverse                    // phares
 #define NUMKEYS   256
 
-static fixed_t default_forwardmove[2] = {0x19, 0x32};
-static fixed_t default_sidemove[2] = {0x18, 0x28};
-fixed_t *forwardmove = default_forwardmove;
+fixed_t forwardmove[2] = {0x19, 0x32};
+fixed_t default_sidemove[2] = {0x18, 0x28};
 fixed_t *sidemove = default_sidemove;
-fixed_t angleturn[3]   = {640, 1280, 320};  // + slow turn
+const fixed_t angleturn[3] = {640, 1280, 320};  // + slow turn
 
 boolean gamekeydown[NUMKEYS];
 int     turnheld;       // for accelerative turning
@@ -391,32 +390,6 @@ static void G_DemoSkipTics(void)
   }
 }
 
-static int RoundSide_Strict(double side)
-{
-  return lround(side * 0.5) * 2; // Even values only.
-}
-
-static int RoundSide_Full(double side)
-{
-  return lround(side);
-}
-
-static int (*RoundSide)(double side) = RoundSide_Full;
-
-void G_UpdateSideMove(void)
-{
-  if (strictmode || (netgame && !solonet))
-  {
-    RoundSide = RoundSide_Strict;
-    sidemove = default_sidemove;
-  }
-  else
-  {
-    RoundSide = RoundSide_Full;
-    sidemove = autostrafe50 ? default_forwardmove : default_sidemove;
-  }
-}
-
 static int CalcControllerForward(int speed)
 {
   const int forward = lroundf(forwardmove[speed] * axes[AXIS_FORWARD] *
@@ -426,14 +399,14 @@ static int CalcControllerForward(int speed)
 
 static int CalcControllerSideTurn(int speed)
 {
-  const int side = RoundSide(forwardmove[speed] * axes[AXIS_TURN] *
+  const int side = G_RoundSide(forwardmove[speed] * axes[AXIS_TURN] *
                              direction[joy_invert_turn]);
   return BETWEEN(-forwardmove[speed], forwardmove[speed], side);
 }
 
 static int CalcControllerSideStrafe(int speed)
 {
-  const int side = RoundSide(forwardmove[speed] * axes[AXIS_STRAFE] *
+  const int side = G_RoundSide(forwardmove[speed] * axes[AXIS_STRAFE] *
                              direction[joy_invert_strafe]);
   return BETWEEN(-sidemove[speed], sidemove[speed], side);
 }
@@ -542,7 +515,7 @@ static int CarryMouseVert(double vert)
 static int CarryMouseSide(double side)
 {
   const double desired = side + prevcarry.side;
-  const int actual = RoundSide(desired);
+  const int actual = G_RoundSide(desired);
   carry.side = desired - actual;
   return actual;
 }
