+++ src/g_input.h
@@ -1,0 +1,26 @@
+//
+// Copyright(C) 2024 ceski
+//
+// This program is free software; you can redistribute it and/or
+// modify it under the terms of the GNU General Public License
+// as published by the Free Software Foundation; either version 2
+// of the License, or (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// DESCRIPTION:
+//      Game Input Utility Functions
+//
+
+#ifndef __G_INPUT__
+#define __G_INPUT__
+
+// Side Movement
+
+extern int (*G_RoundSide)(double side);
+void G_UpdateSideMove(void);
+
+#endif
